requirements.txt
# Core runtime (dashboard)
Flask>=3.0,<4
requests>=2.31,<3

# Optional: calendar assistant (OpenAI / OpenRouter via OpenAI SDK-compatible client)
openai>=1.0,<2

# Optional: Google Calendar integration
google-auth>=2.0,<3
google-auth-oauthlib>=1.0,<2
google-api-python-client>=2.0,<3

# Optional: Blackboard scraper (scripts/scrape_blackboard.py)
python-dotenv>=1.0,<2
selenium>=4.0,<5
webdriver-manager>=4.0,<5

# YAML parsing (used by seed_methods.py for YAML-based library loading)
PyYAML>=6.0,<7

# LangChain + RAG (Adaptive Tutor) + LangGraph (Scholar Deep Agent)
langchain>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-community>=0.3,<0.4
langgraph>=1.0.7,<1.1
chromadb>=0.5,<1
tiktoken>=0.7,<1

# Dev/test
pytest>=8.0,<10
